Integrate HelperFunctionTrait in PageVariable
